♻️ refactor: common.js 생성 후 모듈화 작업
