+++ sign/signin.html
@@ -70,6 +70,6 @@
         </div>
       </article>
     </main>
-    <script src="script/signin.js" type="module"></script>
+    <script type="module" src="script/signin.js"></script>
   </body>
 </html>
