+++ sign/signup.html
@@ -89,6 +89,6 @@
         </div>
       </article>
     </main>
-    <script src="script/signup.js" type="module"></script>
+    <script type="module" src="script/signup.js"></script>
   </body>
 </html>
